feat: Add support for volume rendering
This commit modifies the App component to include support for volume rendering. It enables the usage of the Types.IVolumeViewport interface to retrieve the viewport from the rendering engine. This change allows the application to define a volume in memory and work with volume data.

Refactor the code to work with volume rendering