chore: Update dependencies to latest beta version
This commit updates the dependencies in the package.json file to the latest beta version of the "@cornerstonejs" libraries. Specifically, it updates "@cornerstonejs/core", "@cornerstonejs/dicom-image-loader", "@cornerstonejs/streaming-image-volume-loader", and "@cornerstonejs/tools" to version 2.0.0-beta.16. This update ensures that the project is using the most recent features and bug fixes provided by these libraries.